fix(core-kernel): fix imports from core-kernel utils index file
